First version of Authorization

+++ core/core-impl/src/main/java/org/openbaton/nfvo/core/api/NetworkServiceDescriptorManagement.java
@@ -26,7 +26,6 @@ import org.openbaton.catalogue.mano.record.Status;
 import org.openbaton.catalogue.mano.record.NetworkServiceRecord;
 import org.openbaton.catalogue.mano.descriptor.VirtualNetworkFunctionDescriptor;
 import org.openbaton.catalogue.nfvo.VNFPackage;
-import org.openbaton.catalogue.nfvo.VNFPackage;
 import org.openbaton.catalogue.nfvo.VnfmManagerEndpoint;
 import org.openbaton.exceptions.BadFormatException;
 import org.openbaton.exceptions.CyclicDependenciesException;
@@ -41,6 +40,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Scope;
+import org.springframework.security.oauth2.common.exceptions.UnauthorizedUserException;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.NoResultException;
@@ -89,7 +89,7 @@ public class NetworkServiceDescriptorManagement implements org.openbaton.nfvo.co
      * including any related VNFFGD and VLD.
      */
     @Override
-    public NetworkServiceDescriptor onboard(NetworkServiceDescriptor networkServiceDescriptor) throws NotFoundException, BadFormatException, NetworkServiceIntegrityException, CyclicDependenciesException {
+    public NetworkServiceDescriptor onboard(NetworkServiceDescriptor networkServiceDescriptor, String projectId) throws NotFoundException, BadFormatException, NetworkServiceIntegrityException, CyclicDependenciesException {
 
         log.info("Staring onboarding process for NSD: " + networkServiceDescriptor.getName());
         UrlValidator urlValidator = new UrlValidator();
@@ -98,6 +98,7 @@ public class NetworkServiceDescriptorManagement implements org.openbaton.nfvo.co
         nsdUtils.fetchExistingVnfd(networkServiceDescriptor);
 
         for (VirtualNetworkFunctionDescriptor vnfd : networkServiceDescriptor.getVnfd()) {
+            vnfd.setProjectId(projectId);
             if (vnfd.getLifecycle_event() != null)
                 for (LifecycleEvent event : vnfd.getLifecycle_event()) {
                     if (event == null) {
@@ -113,6 +114,7 @@ public class NetworkServiceDescriptorManagement implements org.openbaton.nfvo.co
                     VNFPackage vnfPackage = new VNFPackage();
                     vnfPackage.setScriptsLink(vnfd.getVnfPackageLocation());
                     vnfPackage.setName(vnfd.getName());
+                    vnfPackage.setProjectId(projectId);
                     vnfPackage = vnfPackageRepository.save(vnfPackage);
                     vnfd.setVnfPackageLocation(vnfPackage.getId());
                 } else { // this is an id pointing to a package already existing
@@ -140,7 +142,7 @@ public class NetworkServiceDescriptorManagement implements org.openbaton.nfvo.co
         }
         log.trace("Creating " + networkServiceDescriptor);
         log.trace("Fetching Data");
-        nsdUtils.fetchVimInstances(networkServiceDescriptor);
+        nsdUtils.fetchVimInstances(networkServiceDescriptor, projectId);
         log.trace("Fetched Data");
 
         log.debug("Checking integrity of NetworkServiceDescriptor");
@@ -151,7 +153,7 @@ public class NetworkServiceDescriptorManagement implements org.openbaton.nfvo.co
         nsdUtils.fetchDependencies(networkServiceDescriptor);
         log.trace("Persisted VNFDependencies");
 
-
+        networkServiceDescriptor.setProjectId(projectId);
         networkServiceDescriptor = nsdRepository.save(networkServiceDescriptor);
         log.info("Created NetworkServiceDescriptor with id " + networkServiceDescriptor.getId());
         return networkServiceDescriptor;
@@ -194,11 +196,14 @@ public class NetworkServiceDescriptorManagement implements org.openbaton.nfvo.co
      * include creating/deleting new VNFFGDs
      * and/or new VLDs.
      *
-     * @param newNsd: the new values to be updated
+     * @param newNsd : the new values to be updated
+     * @param projectId
      */
     @Override
-    public NetworkServiceDescriptor update(NetworkServiceDescriptor newNsd) {
-        return nsdRepository.save(newNsd);
+    public NetworkServiceDescriptor update(NetworkServiceDescriptor newNsd, String projectId) {
+        if (nsdRepository.findFirstById(newNsd.getId()).getProjectId().equals(projectId))
+            return nsdRepository.save(newNsd);
+        throw new UnauthorizedUserException("NSD not under the project chosen, are you trying to hack us? Just kidding, it's a bug :)");
     }
 
     /**
@@ -206,41 +211,55 @@ public class NetworkServiceDescriptorManagement implements org.openbaton.nfvo.co
      *
      * @param vnfd VirtualNetworkFunctionDescriptor to be persisted
      * @param id   of NetworkServiceDescriptor
+     * @param projectId
      * @return the persisted VirtualNetworkFunctionDescriptor
      */
 
-    public VirtualNetworkFunctionDescriptor addVnfd(VirtualNetworkFunctionDescriptor vnfd, String id) {
-        return nsdRepository.addVnfd(vnfd, id);
+    public VirtualNetworkFunctionDescriptor addVnfd(VirtualNetworkFunctionDescriptor vnfd, String id, String projectId) {
+        if (vnfdRepository.findFirstById(vnfd.getId()).getProjectId().equals(projectId))
+            return nsdRepository.addVnfd(vnfd, id) ;
+        throw new UnauthorizedUserException("NSD not under the project chosen, are you trying to hack us? Just kidding, it's a bug :)");
     }
 
     /**
      * Removes the VNFDescriptor with idVnfd from NSD with idNsd
      *
+     * @param nsd
      * @param idNsd  of NSD
      * @param idVnfd of VNFD
      */
     @Override
-    public void deleteVnfDescriptor(String idNsd, String idVnfd) {
+    public void deleteVnfDescriptor(String idNsd, String idVnfd, String projectId) {
         log.info("Removing VnfDescriptor with id: " + idVnfd + " from NSD with id: " + idNsd);
         VirtualNetworkFunctionDescriptor virtualNetworkFunctionDescriptor = vnfdRepository.findFirstById(idVnfd);
+        if (!virtualNetworkFunctionDescriptor.getProjectId().equals(projectId))
+            throw new UnauthorizedUserException("NSD not under the project chosen, are you trying to hack us? Just kidding, it's a bug :)");
         nsdRepository.deleteVnfd(idNsd, idVnfd);
         vnfPackageRepository.delete(virtualNetworkFunctionDescriptor.getVnfPackageLocation());
     }
 
     /**
+     *
      * Returns the VirtualNetworkFunctionDescriptor selected by idVnfd into NSD with idNsd
      *
      * @param idNsd  of NSD
      * @param idVnfd of VirtualNetworkFunctionDescriptor
-     * @return VirtualNetworkFunctionDescriptor
+     * @param projectId
+     * @return
+     * @throws NotFoundException
      */
     @Override
-    public VirtualNetworkFunctionDescriptor getVirtualNetworkFunctionDescriptor(String idNsd, String idVnfd) throws NotFoundException {
+    public VirtualNetworkFunctionDescriptor getVirtualNetworkFunctionDescriptor(String idNsd, String idVnfd, String projectId) throws NotFoundException {
         nsdRepository.exists(idNsd);
-        VirtualNetworkFunctionDescriptor firstById = vnfdRepository.findFirstById(idVnfd);
-        if (firstById == null)
+        VirtualNetworkFunctionDescriptor virtualNetworkFunctionDescriptor = vnfdRepository.findFirstById(idVnfd);
+
+        if (virtualNetworkFunctionDescriptor == null)
             throw new NotFoundException("VirtualNetworkFunctionDescriptor with id " + idVnfd + " doesn't exist");
-        return firstById;
+
+        if (!virtualNetworkFunctionDescriptor.getProjectId().equals(projectId))
+            throw new UnauthorizedUserException("NSD not under the project chosen, are you trying to hack us? Just kidding, it's a bug :)");
+
+        return virtualNetworkFunctionDescriptor;
     }
 
     /**
@@ -249,11 +268,14 @@ public class NetworkServiceDescriptorManagement implements org.openbaton.nfvo.co
      * @param idNsd
      * @param idVfn
      * @param vnfDescriptor
+     * @param projectId
      * @return VirtualNetworkFunctionDescriptor
      */
     @Override
-    public VirtualNetworkFunctionDescriptor updateVNF(String idNsd, String idVfn, VirtualNetworkFunctionDescriptor vnfDescriptor) {
+    public VirtualNetworkFunctionDescriptor updateVNF(String idNsd, String idVfn, VirtualNetworkFunctionDescriptor vnfDescriptor, String projectId) {
         nsdRepository.exists(idNsd);
+        if (!vnfdRepository.findFirstById(vnfDescriptor.getId()).getProjectId().equals(projectId))
+            throw new UnauthorizedUserException("NSD not under the project chosen, are you trying to hack us? Just kidding, it's a bug :)");
         nsdRepository.addVnfd(vnfDescriptor, idNsd);
         return vnfDescriptor;
     }
@@ -263,25 +285,30 @@ public class NetworkServiceDescriptorManagement implements org.openbaton.nfvo.co
      *
      * @param idNsd
      * @param idVnfd
+     * @param projectId
      * @return VNFDependency
      */
     @Override
-    public VNFDependency getVnfDependency(String idNsd, String idVnfd) {
-        nsdRepository.exists(idNsd);
-        return vnfDependencyRepository.findOne(idVnfd);
+    public VNFDependency getVnfDependency(String idNsd, String idVnfd, String projectId) {
+        if (nsdRepository.findFirstById(idNsd).getProjectId().equals(projectId))
+            return vnfDependencyRepository.findOne(idVnfd);
+        throw new UnauthorizedUserException("NSD not under the project chosen, are you trying to hack us? Just kidding, it's a bug :)");
     }
 
     /**
      * Removes the VNFDependency into NSD
-     *
-     * @param idNsd  of NSD
+     *  @param idNsd  of NSD
      * @param idVnfd of VNFD
+     * @param projectId
      */
     @Override
-    public void deleteVNFDependency(String idNsd, String idVnfd) {
+    public void deleteVNFDependency(String idNsd, String idVnfd, String projectId) {
         log.debug("Removing VNFDependency with id: " + idVnfd + " from NSD with id: " + idNsd);
-        nsdRepository.deleteVNFDependency(idNsd, idVnfd);
-        return;
+        if (nsdRepository.findFirstById(idNsd).getProjectId().equals(projectId)) {
+            nsdRepository.deleteVNFDependency(idNsd, idVnfd);
+            return;
+        }
+        throw new UnauthorizedUserException("NSD not under the project chosen, are you trying to hack us? Just kidding, it's a bug :)");
     }
 
     /**
@@ -289,23 +316,29 @@ public class NetworkServiceDescriptorManagement implements org.openbaton.nfvo.co
      *
      * @param idNsd
      * @param vnfDependency
+     * @param projectId
      * @return VNFDependency
      */
     @Override
-    public VNFDependency saveVNFDependency(String idNsd, VNFDependency vnfDependency) {
-        nsdRepository.addVnfDependency(vnfDependency, idNsd);
-        return vnfDependency;
+    public VNFDependency saveVNFDependency(String idNsd, VNFDependency vnfDependency, String projectId) {
+        if (nsdRepository.findFirstById(idNsd).getProjectId().equals(projectId)) {
+            nsdRepository.addVnfDependency(vnfDependency, idNsd);
+            return vnfDependency;
+        }
+        throw new UnauthorizedUserException("NSD not under the project chosen, are you trying to hack us? Just kidding, it's a bug :)");
     }
 
     /**
      * Deletes the PhysicalNetworkFunctionDescriptor from NSD
-     *
-     * @param idNsd of NSD
+     *  @param idNsd of NSD
      * @param idPnf of PhysicalNetworkFunctionDescriptor
+     * @param projectId
      */
     @Override
-    public void deletePhysicalNetworkFunctionDescriptor(String idNsd, String idPnf) {
-        nsdRepository.deletePhysicalNetworkFunctionDescriptor(idNsd, idPnf);
+    public void deletePhysicalNetworkFunctionDescriptor(String idNsd, String idPnf, String projectId) {
+        if (nsdRepository.findFirstById(idNsd).getProjectId().equals(projectId))
+            nsdRepository.deletePhysicalNetworkFunctionDescriptor(idNsd, idPnf);
+        throw new UnauthorizedUserException("NSD not under the project chosen, are you trying to hack us? Just kidding, it's a bug :)");
     }
 
     /**
@@ -313,27 +346,34 @@ public class NetworkServiceDescriptorManagement implements org.openbaton.nfvo.co
      *
      * @param idNsd
      * @param idPnf
+     * @param projectId
      * @return PhysicalNetworkFunctionDescriptor selected
      */
     @Override
-    public PhysicalNetworkFunctionDescriptor getPhysicalNetworkFunctionDescriptor(String idNsd, String idPnf) throws NotFoundException {
-        nsdRepository.exists(idNsd);
-        PhysicalNetworkFunctionDescriptor physicalNetworkFunctionDescriptor = pnfDescriptorRepository.findOne(idPnf);
-        if (physicalNetworkFunctionDescriptor == null)
-            throw new NotFoundException("PhysicalNetworkFunctionDescriptor with id " + idPnf + " doesn't exist");
-        return physicalNetworkFunctionDescriptor;
+    public PhysicalNetworkFunctionDescriptor getPhysicalNetworkFunctionDescriptor(String idNsd, String idPnf, String projectId) throws NotFoundException {
+        if (nsdRepository.findFirstById(idNsd).getProjectId().equals(projectId)) {
+            PhysicalNetworkFunctionDescriptor physicalNetworkFunctionDescriptor = pnfDescriptorRepository.findOne(idPnf);
+            if (physicalNetworkFunctionDescriptor == null)
+                throw new NotFoundException("PhysicalNetworkFunctionDescriptor with id " + idPnf + " doesn't exist");
+            return physicalNetworkFunctionDescriptor;
+        }
+        throw new UnauthorizedUserException("NSD not under the project chosen, are you trying to hack us? Just kidding, it's a bug :)");
     }
 
     /**
      * Add or Update the PhysicalNetworkFunctionDescriptor into NSD
      *
      * @param pDescriptor
-     * @param id
+     * @param idNsd
+     * @param projectId
      * @return PhysicalNetworkFunctionDescriptor
      */
     @Override
-    public PhysicalNetworkFunctionDescriptor addPnfDescriptor(PhysicalNetworkFunctionDescriptor pDescriptor, String id) {
-        return nsdRepository.addPnfDescriptor(pDescriptor, id);
+    public PhysicalNetworkFunctionDescriptor addPnfDescriptor(PhysicalNetworkFunctionDescriptor pDescriptor, String idNsd, String projectId) {
+        if (nsdRepository.findFirstById(idNsd).getProjectId().equals(projectId)) {
+            return nsdRepository.addPnfDescriptor(pDescriptor, idNsd);
+        }
+        throw new UnauthorizedUserException("NSD not under the project chosen, are you trying to hack us? Just kidding, it's a bug :)");
     }
 
     /**
@@ -341,22 +381,34 @@ public class NetworkServiceDescriptorManagement implements org.openbaton.nfvo.co
      *
      * @param id
      * @param security
+     * @param projectId
      * @return Security
      */
     @Override
-    public Security addSecurity(String id, Security security) {
-        return nsdRepository.addSecurity(id, security);
+    public Security addSecurity(String id, Security security, String projectId) {
+        if (nsdRepository.findFirstById(id).getProjectId().equals(projectId)) {
+            return nsdRepository.addSecurity(id, security);
+        }
+        throw new UnauthorizedUserException("NSD not under the project chosen, are you trying to hack us? Just kidding, it's a bug :)");
     }
 
     /**
      * Removes the Secuty with idS from NSD with id
-     *
-     * @param id
+     *  @param idNsd
      * @param idS
+     * @param projectId
      */
     @Override
-    public void deleteSecurty(String id, String idS) {
-        nsdRepository.deleteSecurity(id, idS);
+    public void deleteSecurty(String idNsd, String idS, String projectId) {
+        if (nsdRepository.findFirstById(idNsd).getProjectId().equals(projectId)) {
+            nsdRepository.deleteSecurity(idNsd, idS);
+        }
+        throw new UnauthorizedUserException("NSD not under the project chosen, are you trying to hack us? Just kidding, it's a bug :)");
+    }
+
+    @Override
+    public Iterable<NetworkServiceDescriptor> queryByProjectId(String projectId) {
+        return nsdRepository.findByProjectId(projectId);
     }
 
     /**
@@ -377,8 +429,11 @@ public class NetworkServiceDescriptorManagement implements org.openbaton.nfvo.co
      * related VNFFGD and VLD.
      */
     @Override
-    public NetworkServiceDescriptor query(String id) throws NoResultException {
-        return nsdRepository.findFirstById(id);
+    public NetworkServiceDescriptor query(String id, String projectId) throws NoResultException {
+        NetworkServiceDescriptor networkServiceDescriptor = nsdRepository.findFirstById(id);
+        if (networkServiceDescriptor.getProjectId().equals(projectId))
+            return networkServiceDescriptor;
+        throw new UnauthorizedUserException("NSD not under the project chosen, are you trying to hack us? Just kidding, it's a bug :)");
     }
 
     /**
@@ -388,9 +443,11 @@ public class NetworkServiceDescriptorManagement implements org.openbaton.nfvo.co
      * @param id
      */
     @Override
-    public void delete(String id) throws WrongStatusException{
+    public void delete(String id, String projectId) throws WrongStatusException{
         log.info("Removing NetworkServiceDescriptor with id " + id);
         NetworkServiceDescriptor networkServiceDescriptor = nsdRepository.findFirstById(id);
+        if (!networkServiceDescriptor.getProjectId().equals(projectId))
+            throw new UnauthorizedUserException("NSD not under the project chosen, are you trying to hack us? Just kidding, it's a bug :)");
 
         for (NetworkServiceRecord nsr: nsrRepository.findAll()) {
             if (nsr.getDescriptor_reference().equals(id)) {
@@ -403,7 +460,7 @@ public class NetworkServiceDescriptorManagement implements org.openbaton.nfvo.co
         nsdRepository.delete(networkServiceDescriptor);
         if (cascadeDelete) {
             for (VirtualNetworkFunctionDescriptor virtualNetworkFunctionDescriptor : networkServiceDescriptor.getVnfd()) {
-                virtualNetworkFunctionManagement.delete(virtualNetworkFunctionDescriptor.getId());
+                virtualNetworkFunctionManagement.delete(virtualNetworkFunctionDescriptor.getId(), projectId);
             }
         }
     }
